Cast a size_t for use by printf.
There is no reliable and portable way to printf a size_t.  Yes, I
know about %z in C99, but C99 is currently very far from universal.
Yes, I know about PRIuPTR, but that isn't universal either.

+++ pigz.c
@@ -799,7 +799,8 @@ local void log_dump(void)
         complain("memory leak: %lu allocs of %lu bytes total",
                  mem_track.num, mem_track.size);
     if (mem_track.max)
-        fprintf(stderr, "%lu bytes of memory used\n", mem_track.max);
+        fprintf(stderr, "%lu bytes of memory used\n",
+                (unsigned long)mem_track.max);
 }
 
 /* debugging macro */
